# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Starfield Company Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/{gameId}/players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List players in a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Add a player to a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /games/{gameId}/players/{playerId}:
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Remove a player
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /games/{gameId}/milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones on the earth
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /games/{gameId}/milestones/{milestoneId}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /games/{gameId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks in a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /games/{gameId}/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /games/{gameId}/risks:
    get:
      operationId: listRisks
      tags: [risks]
      summary: List risks (carnivores) in a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Risk"
    post:
      operationId: createRisk
      tags: [risks]
      summary: Create a risk (carnivore)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
  /games/{gameId}/risks/{riskId}:
    patch:
      operationId: updateRisk
      tags: [risks]
      summary: Update a risk
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: riskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
    delete:
      operationId: deleteRisk
      tags: [risks]
      summary: Delete a risk
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: riskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, gameId, name, role, avatarColor, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
        role:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
    PlayerInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        avatarColor:
          type: string
    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        avatarColor:
          type: string
...
    Milestone:
      type: object
      required: [id, gameId, title, status, starsValue, positionX, positionY, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planted, growing, blooming, harvested]
        starsValue:
          type: integer
          description: Value in stars (1-5)
        positionX:
          type: number
          description: X position on earth canvas (0-100)
        positionY:
          type: number
          description: Y position on earth canvas (0-100)
        targetDate:
          type: ["string", "null"]
        taskCount:
          type: ["integer", "null"]
        completedTaskCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    MilestoneInput:
      type: object
      required: [title, starsValue]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        starsValue:
          type: integer
          minimum: 1
          maximum: 5
        positionX:
          type: number
        positionY:
          type: number
        targetDate:
          type: string
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planted, growing, blooming, harvested]
        starsValue:
          type: integer
        positionX:
          type: number
        positionY:
          type: number
        targetDate:
          type: string
    Task:
      type: object
      required: [id, gameId, title, status, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        milestoneId:
          type: ["integer", "null"]
        assignedPlayerId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [plan, doing, check, next_plan]
          description: "Kanban stages: plan=Plan, doing=Doing, check=Review KPI + Residual Close, next_plan=Next Plan"
        priority:
          type: string
          enum: [low, medium, high, critical]
        kpiImpact:
          type: ["string", "null"]
          description: Which KPI this task impacts
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        milestoneId:
          type: ["integer", "null"]
        assignedPlayerId:
          type: ["integer", "null"]
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        kpiImpact:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        milestoneId:
          type: ["integer", "null"]
        assignedPlayerId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [plan, doing, check, next_plan]
        priority:
          type: string
          enum: [low, medium, high, critical]
        kpiImpact:
          type: string
        dueDate:
          type: string
    Risk:
      type: object
      required: [id, gameId, name, severity, status, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
          description: The carnivore label
        description:
          type: ["string", "null"]
        severity:
          type: string
          enum: [lurking, stalking, charging, tamed]
        likelihood:
          type: integer
          description: 1-5 likelihood stars
        impact:
          type: integer
          description: 1-5 impact stars
        status:
          type: string
          enum: [active, mitigated, eliminated]
        mitigationPlan:
          type: ["string", "null"]
        positionX:
          type: ["number", "null"]
        positionY:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    RiskInput:
      type: object
      required: [name, severity]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        severity:
          type: string
          enum: [lurking, stalking, charging, tamed]
        likelihood:
          type: integer
          minimum: 1
          maximum: 5
        impact:
          type: integer
          minimum: 1
          maximum: 5
        mitigationPlan:
          type: string
        positionX:
          type: number
        positionY:
          type: number
    RiskUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [lurking, stalking, charging, tamed]
        likelihood:
          type: integer
        impact:
          type: integer
        status:
          type: string
          enum: [active, mitigated, eliminated]
        mitigationPlan:
          type: string
        positionX:
          type: number
        positionY:
          type: number
...