# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Starfield Company Game API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  gameId: number;
  /** @nullable */
  milestoneId?: number | null;
  /** @nullable */
  assignedPlayerId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** Kanban stages: plan=Plan, doing=Doing, check=Review KPI + Residual Close, next_plan=Next Plan */
  status: TaskStatus;
  priority?: TaskPriority;
  /**
     * Which KPI this task impacts
     * @nullable
     */
  kpiImpact?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: Date;
}